ScenesReact: Make useVariableValues generic and rename VariableSelect to VariableControl
