+++ packages/scenes-app/src/react-demo/DynamicVariablesPage.tsx
@@ -1,4 +1,4 @@
-import { CustomVariable, VariableSelect } from '@grafana/scenes-react';
+import { CustomVariable, VariableControl } from '@grafana/scenes-react';
 import { Stack } from '@grafana/ui';
 import React from 'react';
 import { PageWrapper } from './PageWrapper';
@@ -11,7 +11,7 @@ export function DynamicVariablesPage() {
       <CustomVariable name="job" query="A, B, C" initialValue="A">
         <Stack direction="column">
           <Stack>
-            <VariableSelect name="job" />
+            <VariableControl name="job" />
           </Stack>
           <DemoVizLayout>
             <PlainGraphWithRandomWalk title={'Testing job = $job'} queryAlias="job = $job" />
