+++ packages/scenes-app/src/react-demo/PageWrapper.tsx
@@ -1,6 +1,6 @@
 import { NavModelItem } from '@grafana/data';
 import { PluginPage } from '@grafana/runtime';
-import { BreadcrumbContext, RefreshPicker, TimeRangePicker, VariableSelect } from '@grafana/scenes-react';
+import { BreadcrumbContext, RefreshPicker, TimeRangePicker, VariableControl } from '@grafana/scenes-react';
 import React, { useContext } from 'react';
 
 export interface Props {
@@ -35,7 +35,7 @@ export function PageWrapper({ title, subTitle, children }: Props) {
 function PageActions() {
   return (
     <>
-      <VariableSelect name="env" />
+      <VariableControl name="env" />
       <TimeRangePicker />
       <RefreshPicker withText={true} />
     </>
