+++ packages/scenes-react/src/index.ts
@@ -1,7 +1,7 @@
 export { useQueryRunner } from './hooks/useQueryRunner';
 export * from './contexts/SceneContextProvider';
 export { TimeRangePicker } from './components/TimeRangePicker';
-export { VariableSelect } from './components/VariableSelect';
+export { VariableControl } from './components/VariableControl';
 export { VizPanel } from './components/VizPanel';
 export { RefreshPicker } from './components/RefreshPicker';
 export { CustomVariable } from './variables/CustomVariable';
